# Generate `__version__` at build to avoid slow `importlib.metadata` import