--- a/pyproject.toml
+++ b/pyproject.toml
@@ -53,6 +53,9 @@ urls.Source = "https://github.com/pylast/pylast"
 [tool.hatch]
 version.source = "vcs"
 
+[tool.hatch.build.hooks.vcs]
+version-file = "src/pylast/_version.py"
+
 [tool.hatch.version.raw-options]
 local_scheme = "no-local-version"