Like https://github.com/hugovk/tinytext/pull/195.

With Python 3.13.0rc2:

```sh
python -X importtime -c "import pylast" 2> import.log && tuna import.log
```

# `main`

<img width="1582" alt="image" src="https://github.com/user-attachments/assets/d3b95572-74f2-4927-a1d7-f7f358591dca">

# PR

<img width="1582" alt="image" src="https://github.com/user-attachments/assets/91fe7a6b-1f66-4e36-8b09-05a4c629a5fc">



---

httpx remains the slowest import, but it's very likely it is needed when running pylast. We could replace it with something faster to import like urllib3, but that's for another time.

```console
❯ hyperfine --warmup 32  'python3 -c "import httpx"' 'python3 -c "import urllib3"'
Benchmark 1: python3 -c "import httpx"
  Time (mean ± σ):     146.8 ms ± 114.8 ms    [User: 104.7 ms, System: 18.3 ms]
  Range (min … max):   119.2 ms … 685.5 ms    24 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Benchmark 2: python3 -c "import urllib3"
  Time (mean ± σ):      47.8 ms ±   1.5 ms    [User: 38.9 ms, System: 7.8 ms]
  Range (min … max):    46.5 ms …  56.7 ms    60 runs

  Warning: Statistical outliers were detected. Consider re-running this benchmark on a quiet system without any interferences from other programs. It might help to use the '--warmup' or '--prepare' options.

Summary
  python3 -c "import urllib3" ran
    3.07 ± 2.41 times faster than python3 -c "import httpx"
```

urllib3 also has no dependencies (httpx has 7) and is a bit smaller (13 MB) than httpx (16 MB), so is a bit quicker to install:

<details>
<summary>Details</summary>

```console
❯ python -m venv urllib3

❯ time urllib3/bin/pip install urllib3 --no-cache-dir
Collecting urllib3
  Downloading urllib3-2.2.2-py3-none-any.whl.metadata (6.4 kB)
Downloading urllib3-2.2.2-py3-none-any.whl (121 kB)
Installing collected packages: urllib3
Successfully installed urllib3-2.2.2
urllib3/bin/pip install urllib3 --no-cache-dir  0.28s user 0.10s system 38% cpu 0.980 total

❯ du -sh urllib3
 13Murllib3

❯ python -m venv httpx
Found existing alias for "python". You should use: "p"

❯ time httpx/bin/pip install httpx --no-cache-dir
Collecting httpx
  Downloading httpx-0.27.2-py3-none-any.whl.metadata (7.1 kB)
Collecting anyio (from httpx)
  Downloading anyio-4.4.0-py3-none-any.whl.metadata (4.6 kB)
Collecting certifi (from httpx)
  Downloading certifi-2024.8.30-py3-none-any.whl.metadata (2.2 kB)
Collecting httpcore==1.* (from httpx)
  Downloading httpcore-1.0.5-py3-none-any.whl.metadata (20 kB)
Collecting idna (from httpx)
  Downloading idna-3.8-py3-none-any.whl.metadata (9.9 kB)
Collecting sniffio (from httpx)
  Downloading sniffio-1.3.1-py3-none-any.whl.metadata (3.9 kB)
Collecting h11<0.15,>=0.13 (from httpcore==1.*->httpx)
  Downloading h11-0.14.0-py3-none-any.whl.metadata (8.2 kB)
Downloading httpx-0.27.2-py3-none-any.whl (76 kB)
Downloading httpcore-1.0.5-py3-none-any.whl (77 kB)
Downloading anyio-4.4.0-py3-none-any.whl (86 kB)
Downloading idna-3.8-py3-none-any.whl (66 kB)
Downloading sniffio-1.3.1-py3-none-any.whl (10 kB)
Downloading certifi-2024.8.30-py3-none-any.whl (167 kB)
Downloading h11-0.14.0-py3-none-any.whl (58 kB)
Installing collected packages: sniffio, idna, h11, certifi, httpcore, anyio, httpx
Successfully installed anyio-4.4.0 certifi-2024.8.30 h11-0.14.0 httpcore-1.0.5 httpx-0.27.2 idna-3.8 sniffio-1.3.1
httpx/bin/pip install httpx --no-cache-dir  0.45s user 0.11s system 47% cpu 1.192 total

❯ du -sh httpx
 16Mhttpx
```


</details>